Rename view components to be multi-word
To comply with https://eslint.vuejs.org/rules/multi-word-component-names
