+++ src/router.ts
@@ -1,7 +1,7 @@
 import {createRouter, createWebHistory} from 'vue-router'
-import Home from './views/Home/Home.vue'
-import Locations from './views/Locations/Locations.vue'
-import Settings from './views/Settings/Settings.vue'
+import Home from './views/Home/HomeView.vue'
+import Locations from './views/Locations/LocationsView.vue'
+import Settings from './views/Settings/SettingsView.vue'
 
 const routes = [
   {
