+++ src/views/Locations/LocationsView.vue
