+++ src/views/Settings/SettingsView.vue
